Wine Quality Project
The wines data set contains data on 1599 wines columns and 12 variables from around 2009. I started by asking some questions,then explored with the dataset,created several plots to answer my questions. First of all,I made a assumption that factors might affect the quality of the wine. For example, pH was negatively correlated to volatile acidity, which makes sense.

I created a linear model to attempt to predict red wine qualities, which was accurate for average wines but extremely inaccurate for bad/excellent wines; it either over predicted bad wines and underpredicted the good ones.

Alcohol appeared to be the number one factor in determining an excellent wine. However, Citric acid and sulphates had to be in specific amounts in order for alcohol to take over. Volatile acidity made a wine bad in large amounts.